+++ workers/worker_parser_loop.py
@@ -3,7 +3,7 @@ import logging
 
 from sentry_sdk import capture_exception
 
-from runner.runner_async import runner as runner_async_func
+from runner.runner import runner
 
 
 logger = logging.getLogger(__name__)
@@ -17,7 +17,7 @@ class ParserLoopWorker:
             # waiting before run to allow other services some time to start
             await asyncio.sleep(3 * 60)
             try:
-                await runner_async_func()
+                await runner()
             except Exception as e:
                 capture_exception(e)
                 logger.error(f"ERROR: {e}")
